# docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  model.Pet:
    properties:
      age:
        type: integer
      breed:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
paths:
  /pets:
    get:
      description: Get a list of all pets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Pet'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List all pets
      tags:
      - pets
    post:
      consumes:
      - application/json
      description: Add a new pet to the system
      parameters:
      - description: New pet
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/model.Pet'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Pet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new pet
      tags:
      - pets
  /pets/{id}:
    delete:
      description: Delete a pet by ID
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete a pet
      tags:
      - pets
    get:
      description: Get details of a pet by ID
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Pet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a pet by ID
      tags:
      - pets
  /pets/type:
    get:
      description: Get a list of pets by type
      parameters:
      - description: Pet Type
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Pet'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get pets by type
      tags:
      - pets
swagger: "2.0"